feat: card component 수정 및 react-router 적용
