+++ src/styles/LinkCard.module.css
@@ -3,6 +3,7 @@
   border-radius: 15px;
   box-shadow: 0px 5px 25px 0px #00000014;
   overflow: hidden;
+  position: relative;
 }
 .linkGridItem:hover {
   transform: scale(1.1);
@@ -11,6 +12,18 @@
 .itemImg img {
   width: 100%;
   height: 250px;
+}
+
+.starBtn {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+}
+
+.flexBox {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
 }
 
 .linkGridItem .itemContext {
